+++ src/pages/HomePage/HomePage.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { Category, ProductList, FeedBack, BlogPost } from '@components';
+import { Category, ProductList, FeedBack, BlogPost, MenuBar } from '@components';
 
 export const HomePage: React.FC = () => {
   return (
     <>
+      <MenuBar />
       <Category />
       <ProductList />
       <FeedBack />
