+++ src/pages/LinkPage/LinkPage.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Switch, Route as MyRouter } from 'react-router-dom';
-import { HomePage } from '@pages';
-import { ProductDetail } from '@components';
-import { CartPage } from '@components';
+import { HomePage, CartPage, ProductDetailPage } from '@pages';
 
 export const LinkPage: React.FC = () => {
   return (
@@ -15,7 +13,7 @@ export const LinkPage: React.FC = () => {
         // }}
         component={HomePage}
       />
-      <MyRouter path="/product" component={ProductDetail} />
+      <MyRouter path="/product" component={ProductDetailPage} />
       <MyRouter path="/cart" component={CartPage} />
     </Switch>
   );
