+++ src/pages/ProductDetailPage/ProductDetailPage.css
@@ -1,4 +1,19 @@
-
+.product__detail-shipping-btn {
+    border-radius: 1.2rem;
+    padding: 0 .8rem;
+    position: absolute;
+    z-index: 10;
+    color: var(--green);
+    background: var(--gray-bg3);
+}
+.product__detail-percent-btn {
+    border-radius: 1.2rem;
+    padding: 0 .8rem;
+    position: absolute;
+    z-index: 10;
+    color: var(--white);
+    background: var(--red);
+}
 .tab__btn {
     padding: 1.6rem 1.6rem 1.6rem 0 !important;
     width: max-content !important;
