+++ src/pages/ProductDetailPage/components/FavoriteComponent.tsx
