+++ src/pages/ProductDetailPage/components/InfoDetail.tsx
