+++ src/pages/ProductDetailPage/components/ProportionComponent.tsx
