+++ src/pages/index.ts
@@ -1,3 +1,5 @@
 export { default as RootPage } from './Root/RootPage';
 export * from './LinkPage/LinkPage';
 export * from './HomePage/HomePage';
+export * from './ProductDetailPage/ProductDetailPage';
+export * from './CartPage/CartPage';
